Moved AVI dumping into its own folder, and split logic out across multiple files.
