+++ source/application/application.vcxproj
@@ -136,7 +136,10 @@
     </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="AVIDumper.cpp" />
+    <ClCompile Include="AVIDumping\AudioConverterStream.cpp" />
+    <ClCompile Include="AVIDumping\AVIDumper.cpp" />
+    <ClCompile Include="AVIDumping\WaitableBool.cpp" />
+    <ClCompile Include="AVIDumping\WaveFormat.cpp" />
     <ClCompile Include="Config.cpp" />
     <ClCompile Include="CPUinfo.cpp" />
     <ClCompile Include="CustomDLGs.cpp" />
@@ -160,7 +163,10 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\shared\input.h" />
-    <ClInclude Include="AVIDumper.h" />
+    <ClInclude Include="AVIDumping\AudioConverterStream.h" />
+    <ClInclude Include="AVIDumping\AVIDumper.h" />
+    <ClInclude Include="AVIDumping\WaitableBool.h" />
+    <ClInclude Include="AVIDumping\WaveFormat.h" />
     <ClInclude Include="Config.h" />
     <ClInclude Include="CPUinfo.h" />
     <ClInclude Include="CustomDLGs.h" />
